Create game steps
