--- Mobile/Android/app/src/main/res/layout/fragment_create_game.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-
-    <androidx.appcompat.widget.LinearLayoutCompat
-        android:id="@+id/clueControls"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_margin="10dp"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toTopOf="@id/create_game_button"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-
-        <ImageView
-            android:id="@+id/imagePreview"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:scaleType="fitCenter"
-            android:tag="0"
-            android:src="@drawable/magnifying_glass" />
-
-        <androidx.appcompat.widget.LinearLayoutCompat
-            android:id="@+id/imageControls"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:minHeight="5dp"
-            android:orientation="horizontal">
-
-            <Button
-                android:id="@+id/upload_image_button"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:enabled="true"
-                android:text="@string/upload_image" />
-
-            <Button
-                android:id="@+id/take_photo_button"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:enabled="true"
-                android:text="@string/take_photo" />
-
-        </androidx.appcompat.widget.LinearLayoutCompat>
-
-        <EditText
-            android:id="@+id/game_clue_edit_text"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="2"
-            android:enabled="true"
-            android:hint="@string/game_give_clue"
-            android:inputType="textMultiLine"
-            android:textAlignment="center"
-            />
-
-    </androidx.appcompat.widget.LinearLayoutCompat>
-
-    <Button
-        android:id="@+id/create_game_button"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:enabled="true"
-        android:text="@string/create"
-        android:layout_margin="10dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/clueControls"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
