+++ Mobile/Android/app/src/main/res/layout/fragment_create_game_step_add_clue.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.appcompat.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:layout_margin="10dp">
+
+    <androidx.appcompat.widget.AppCompatTextView
+        android:layout_width="match_parent"
+        android:layout_height="48dp"
+        android:layout_weight="0"
+        android:text="@string/enter_clue"
+        android:textAlignment="center"
+        android:textSize="@dimen/title" />
+
+    <androidx.appcompat.widget.AppCompatEditText
+        android:id="@+id/game_clue_edit_text"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:enabled="true"
+        android:hint="@string/game_give_clue"
+        android:inputType="textMultiLine"
+        android:textAlignment="center" />
+
+    <androidx.appcompat.widget.AppCompatButton
+        android:id="@+id/next_button"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:layout_weight="0"
+        android:enabled="true"
+        android:textAlignment="center"
+        android:text="@string/next" />
+
+</androidx.appcompat.widget.LinearLayoutCompat>
